+++ backtrace-library/src/main/java/backtraceio/library/models/json/BacktraceReport.java
@@ -2,11 +2,11 @@ package backtraceio.library.models.json;
 
 import android.content.Context;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
-import java.lang.Exception;
 
 import backtraceio.library.common.BacktraceTimeHelper;
 import backtraceio.library.common.CollectionUtils;
@@ -14,7 +14,6 @@ import backtraceio.library.models.BacktraceAttributeConsts;
 import backtraceio.library.models.BacktraceData;
 import backtraceio.library.models.BacktraceStackFrame;
 import backtraceio.library.models.BacktraceStackTrace;
-import backtraceio.library.models.UnhandledThrowableWrapper;
 
 /**
  * Captured application error
@@ -22,46 +21,42 @@ import backtraceio.library.models.UnhandledThrowableWrapper;
 public class BacktraceReport {
 
     /**
+     * Reference to the original exception. This field is internally used by the library.
+     */
+    private final transient Throwable originalException;
+    /**
      * 16 bytes of randomness in human readable UUID format
      * server will reject request if uuid is already found
      */
     public UUID uuid = UUID.randomUUID();
-
     /**
      * UTC timestamp in seconds
      */
     public long timestamp = BacktraceTimeHelper.getTimestampSeconds();
-
     /**
      * Get information about report type. If value is true the BacktraceReport has an error
      */
     public Boolean exceptionTypeReport = false;
-
     /**
      * Get a report classification
      */
     public String classifier = "";
-
     /**
      * Get an report attributes
      */
-    public Map<String, Object> attributes;
-
+    public Map<String, Object> attributes = new HashMap<>();
     /**
      * Get a custom client message
      */
     public String message;
-
     /**
      * Get a report exception
      */
     public Exception exception;
-
     /**
      * Get all paths to attachments
      */
-    public List<String> attachmentPaths;
-
+    public List<String> attachmentPaths = new ArrayList<>();
     /**
      * Current report exception stack
      */
@@ -133,7 +128,7 @@ public class BacktraceReport {
      * @param exception current exception
      */
     public BacktraceReport(
-            Exception exception) {
+            Throwable exception) {
         this(exception, null, null);
     }
 
@@ -145,7 +140,7 @@ public class BacktraceReport {
      * @param attributes additional information about application state
      */
     public BacktraceReport(
-            Exception exception,
+            Throwable exception,
             Map<String, Object> attributes) {
         this(exception, attributes, null);
     }
@@ -167,24 +162,22 @@ public class BacktraceReport {
      * Create new instance of Backtrace report to send a report
      * with application exception, attributes and attachments
      *
-     * @param exception       current exception
+     * @param throwable       current throwable
      * @param attributes      additional information about application state
      * @param attachmentPaths path to all report attachments
      */
-    public BacktraceReport(
-            Exception exception,
-            Map<String, Object> attributes,
-            List<String> attachmentPaths) {
-
+    public BacktraceReport(Throwable throwable, Map<String, Object> attributes, List<String> attachmentPaths) {
         this.attributes = CollectionUtils.copyMap(attributes);
         this.attachmentPaths = CollectionUtils.copyList(attachmentPaths);
-        this.exception = this.prepareException(exception);
-        this.exceptionTypeReport = exception != null;
+        this.originalException = throwable;
+
+        if (throwable != null) {
+            this.exception = this.prepareException(throwable);
+            this.classifier = this.getExceptionClassifier(throwable);
+            this.exceptionTypeReport = true;
+        }
         this.diagnosticStack = new BacktraceStackTrace(exception).getStackFrames();
 
-        if (this.exceptionTypeReport && exception != null) {
-            this.classifier = getExceptionClassifier(exception);
-        }
         this.setDefaultErrorTypeAttribute();
     }
 
@@ -200,50 +193,11 @@ public class BacktraceReport {
         this.classifier = classifier;
         this.attributes = attributes;
         this.message = message;
+        this.originalException = exception;
         this.exception = exception;
         this.attachmentPaths = attachmentPaths;
         this.diagnosticStack = diagnosticStack;
     }
-
-    /**
-     * To avoid serialization issues with custom exceptions, our goal is to always
-     * prepare exception in a way potential serialization won't break it
-     *
-     * @param exception captured client-side exception
-     */
-    private Exception prepareException(Exception exception) {
-        if (exception == null) {
-            return null;
-        }
-        Exception reportException = new Exception(exception.getMessage());
-        reportException.setStackTrace(exception.getStackTrace());
-
-        return reportException;
-    }
-
-    public String getExceptionClassifier(Exception exception) {
-        if (exception instanceof UnhandledThrowableWrapper) {
-            return ((UnhandledThrowableWrapper) exception).getClassifier();
-        }
-        return exception.getClass().getCanonicalName();
-    }
-
-    /**
-     * Sets error.type attribute depends on the type of the report
-     */
-    private void setDefaultErrorTypeAttribute() {
-        if (attributes.containsKey(BacktraceAttributeConsts.ErrorType)) {
-            // error type already set
-            return;
-        }
-
-        attributes.put(
-                BacktraceAttributeConsts.ErrorType,
-                this.exceptionTypeReport
-                        ? BacktraceAttributeConsts.HandledExceptionAttributeType
-                        : BacktraceAttributeConsts.MessageAttributeType);
-    }
-
     /**
      * Concat two dictionaries with attributes
      *
@@ -262,6 +216,69 @@ public class BacktraceReport {
         return reportAttributes;
     }
 
+    /**
+     * Returns report exception - original exception if possible or serializable exception if
+     * the report was loaded from JSON.
+     *
+     * @return throwable object
+     */
+    public Throwable getException() {
+        return this.originalException != null ? this.originalException : this.exception;
+    }
+
+//    public BacktraceData toBacktraceData(Context context, Map<String, Object> clientAttributes) {
+//        return toBacktraceData(context, clientAttributes, false);
+//    }
+//
+//    public BacktraceData toBacktraceData(Context context, Map<String, Object> clientAttributes, boolean isProguardEnabled) {
+//        BacktraceData backtraceData = new BacktraceData(context, this, clientAttributes);
+//        backtraceData.symbolication = isProguardEnabled ? "proguard" : null;
+//        return backtraceData;
+//    }
+
+    private String getExceptionClassifier(Throwable exception) {
+        return exception.getClass().getCanonicalName();
+    }
+
+    /**
+     * To avoid serialization issues with custom exceptions, our goal is to always
+     * prepare exception in a way potential serialization won't break it
+     *
+     * @param exception captured client-side exception
+     */
+    private Exception prepareException(Throwable exception) {
+        if (exception == null) {
+            return null;
+        }
+        Exception reportException = new Exception(exception.getMessage());
+        reportException.setStackTrace(exception.getStackTrace());
+
+        return reportException;
+    }
+
+//    public String getExceptionClassifier(Exception exception) {
+//        if (exception instanceof UnhandledThrowableWrapper) {
+//            return ((UnhandledThrowableWrapper) exception).getClassifier();
+//        }
+//        return exception.getClass().getCanonicalName();
+//    }
+
+    /**
+     * Sets error.type attribute depends on the type of the report
+     */
+    private void setDefaultErrorTypeAttribute() {
+        if (attributes.containsKey(BacktraceAttributeConsts.ErrorType)) {
+            // error type already set
+            return;
+        }
+
+        attributes.put(
+                BacktraceAttributeConsts.ErrorType,
+                this.exceptionTypeReport
+                        ? BacktraceAttributeConsts.HandledExceptionAttributeType
+                        : BacktraceAttributeConsts.MessageAttributeType);
+    }
+
     public BacktraceData toBacktraceData(Context context, Map<String, Object> clientAttributes) {
         return toBacktraceData(context, clientAttributes, false);
     }
@@ -271,3 +288,4 @@ public class BacktraceReport {
         return new BacktraceData.Builder(this).setAttributes(context, clientAttributes).setSymbolication(symbolication).build();
     }
 }
+